Local changes

+++ src/segment.hpp
@@ -39,7 +39,7 @@ class Segment
 		Segment& operator*=(TValue s);
 		Segment& operator/=(TValue s);
 
-		bool intersection( const Segment& s1,double& x, double& y);
+		bool intersection( const Segment& s1,float& x, float& y);
 		bool intersection( const Segment& s1);
